Warning and error fixes.
svn path=/branches/ros-amd64-bringup/; revision=34906

+++ reactos/base/applications/network/tracert/tracert.c
@@ -388,7 +388,7 @@ DecodeResponse(PAPPINFO pInfo)
     switch (IcmpHdr->icmpheader.type)
     {
            case TTL_EXCEEDED :
-                _tprintf(_T("%3ld ms"), (ULONG)((pInfo->lTimeEnd - pInfo->lTimeStart) / pInfo->TicksPerMs.QuadPart));
+                _tprintf(_T("%3I64d ms"), (pInfo->lTimeEnd - pInfo->lTimeStart) / pInfo->TicksPerMs.QuadPart);
                 return 0;
 
            case ECHO_REPLY :
@@ -399,7 +399,7 @@ DecodeResponse(PAPPINFO pInfo)
                     DebugPrint(_T("Rouge packet: header id %d, process id  %d"), IcmpHdr->icmpheader.id, GetCurrentProcessId());
                     return -1;
                 }
-                _tprintf(_T("%3ld ms"), (ULONG)((pInfo->lTimeEnd - pInfo->lTimeStart) / pInfo->TicksPerMs.QuadPart));
+                _tprintf(_T("%3I64d ms"), (pInfo->lTimeEnd - pInfo->lTimeStart) / pInfo->TicksPerMs.QuadPart);
                 return 1;
 
            case DEST_UNREACHABLE :
